+++ src/domains/data-provider-v2/providers/globo-esporte/index.ts
@@ -1,0 +1,10 @@
+import { IApiProviderV2 } from '../../interface';
+import { GloboEsporteMatches } from './globo-esport-matches';
+import { GloboEsporteTeams } from './globo-esporte-teams';
+import { GloboEsporteTournament } from './globo-esporte-tournament';
+
+export const ApiProviderGloboEsporte = {
+  tournament: GloboEsporteTournament,
+  teams: GloboEsporteTeams,
+  matches: GloboEsporteMatches,
+} satisfies IApiProviderV2;
