+++ src/domains/match/controllers/match-controller.ts
@@ -1,4 +1,4 @@
-import { ACTIVE_PROVIDER } from '@/domains/data-providers';
+import { ACTIVE_API_PROVIDER } from '@/domains/data-provider-v2';
 import { ErrorMapper } from '@/domains/match/error-handling/mapper';
 import { T_Match } from '@/domains/match/schema';
 import { GlobalErrorMapper } from '@/domains/shared/error-handling/mapper';
@@ -47,7 +47,7 @@ async function getMatchesByTournament(req: Request, res: Response) {
       .where(
         and(
           eq(T_Match.tournamentId, tournamentId),
-          eq(T_Match.provider, ACTIVE_PROVIDER),
+          eq(T_Match.provider, ACTIVE_API_PROVIDER),
           eq(T_Match.roundId, round)
         )
       );
